Refactor file uploader component and extract image upload logic into a new ImageUploader component
